requirements.txt
twilio>=8.0.0
python-dotenv>=1.0.0
azure-communication-email>=1.0.0

